pyproject.toml
[project]
name = "ontology-skill"
version = "0.1.0"
description = "Ontology engineering workspace with Claude Code skills for programmatic ontology development"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    # Core ontology tools
    "oaklib>=0.6.0",
    "rdflib>=7.0.0",
    "pyshacl>=0.26.0",
    "owlready2>=0.46",
    # Mapping
    "sssom>=0.4.0",
    "curies>=0.7.0",
    "prefixmaps>=0.2.0",
    # Schema / modeling
    "linkml>=1.8.0",
    "linkml-runtime>=1.8.0",
    # SPARQL
    "SPARQLWrapper>=2.0.0",
    # Data handling
    "pandas>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    # HTTP
    "httpx>=0.27.0",
    "trafilatura>=2.0,<3",
    "pyoxigraph>=0.5.5",
]

[dependency-groups]
dev = [
    # Linting & formatting
    "ruff>=0.9.0",
    # Type checking
    "mypy>=1.14.0",
    "pandas-stubs>=2.0",
    "types-PyYAML>=6.0",
    "types-requests>=2.31",
    # Testing
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "respx>=0.21.0",
    "pip-audit>=2.8.0",
    # Pre-commit
    "pre-commit>=4.0",
    "pylode>=3.2.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ---------------------------------------------------------------------------
# Ruff - linting and formatting
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "scripts"]
extend-exclude = [
    "**/notebooks/",          # Jupyter notebooks follow different conventions
    "**/.checkpoints/",       # historical ontology snapshots; not live code
    "**/.cache/",             # tool caches
]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "ASYNC", # flake8-async
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "ISC",   # flake8-implicit-str-concat
    "ICN",   # flake8-import-conventions
    "PIE",   # flake8-pie
    "PT",    # flake8-pytest-style
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SLF",   # flake8-self
    "SIM",   # flake8-simplify
    "TID",   # flake8-tidy-imports
    "TCH",   # flake8-type-checking
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    "ERA",   # eradicate
    "PL",    # pylint
    "TRY",   # tryceratops
    "FLY",   # flynt
    "PERF",  # perflint
    "RUF",   # ruff-specific
    "S",     # flake8-bandit
]
ignore = [
    "COM812",   # trailing comma (conflicts with formatter)
    "ISC001",   # single-line implicit string concat (conflicts with formatter)
    "TRY003",   # long exception messages (too noisy)
    "PLR0913",  # too many function arguments (ontology APIs need them)
    "PLR2004",  # magic values (too noisy for data processing)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",     # unused arguments common in fixtures
    "SLF001",  # private member access in tests is fine
    "PLR2004", # magic values in assertions
    "PLC0415", # imports inside test functions are intentional
    "S101",    # asserts are normal in pytest tests
]
"ontologies/**/tests/**/*.py" = [
    "ARG",     # unused arguments common in fixtures
    "SLF001",  # private member access in tests is fine
    "PLR2004", # magic values in assertions
    "PLC0415", # imports inside test functions are intentional
    "S101",    # asserts are normal in pytest tests
]
"scripts/**/*.py" = [
    "T201",    # print statements OK in scripts
]
"ontologies/**/scripts/**/*.py" = [
    "T201",    # print statements OK in ontology build scripts
    "PLR0912", # data-processing functions have many branches by design
    "PLR0915", # data-loading build functions have many statements by design
    "PLC0415", # deferred/optional imports in CLI entrypoints are intentional
    "S311",    # seeded PRNG is fine for stratified sampling (not crypto)
]

[tool.ruff.lint.isort]
known-first-party = ["ontology_skill"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ---------------------------------------------------------------------------
# mypy - type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false
no_implicit_optional = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "owlready2.*",
    "owlapy.*",
    "pyshacl.*",
    "sssom.*",
    "oaklib.*",
    "linkml.*",
    "linkml_runtime.*",
    "SPARQLWrapper.*",
    "prefixmaps.*",
    "gilda.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests", "ontologies"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
skip_empty = true
